pyproject.toml
[project]
name = "studywidgetapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mediapipe>=0.10.32",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=11.0.0",
    "pyside6>=6.10.2",
    "questionary>=2.1.1",
    "rich>=14.1.0",
    "torch>=2.10.0",
    "timm>=0.9.0",           # required by Grounding DINO backbone
    "transformers>=5.3.0",
    "ultralytics>=8.3", #YOLO V11
]

[tool.uv]
# Restricts dependency resolution to only these platforms.
# This is needed because mediapipe only provides pre-built wheels for
# these specific OS + architecture combinations — without this, uv may
# try to resolve for unsupported platforms (e.g. 32-bit Windows) and fail.
required-environments = [
    "sys_platform == 'win32' and platform_machine == 'AMD64'",    # Windows 64-bit
    "sys_platform == 'darwin' and platform_machine == 'arm64'",   # macOS Apple Silicon
    "sys_platform == 'linux' and platform_machine == 'x86_64'",   # Linux 64-bit
]
